add styles

+++ backend/upload/upload_delete.php
@@ -4,11 +4,14 @@ session_start();
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
   require_once "../connect.php";
   
-  if (isset($_POST['uuid'])) {
+  if (isset($_POST['uuid']) && isset($_POST['type'])) {
     $uuid = $_POST['uuid'];
+    $type = $_POST['type'];
+    $query_str = "DELETE FROM $type WHERE uuid=?;";
+    
     unlink('../uploads/'. $uuid . ".png");
     try {
-      $pdo->prepare("DELETE FROM products WHERE uuid=?;")->execute([$uuid]);
+      $pdo->prepare($query_str)->execute([$uuid]);
     } catch (PDOException $e) {
       header("Location: ../../frontend/routes/product-info.php");
     }
